--- src/main/java/com/api/ttoklip/global/security/auth/dto/LoginResponse.java
@@ -1,7 +1,0 @@
-package com.api.ttoklip.global.security.auth.dto;
-
-import lombok.Builder;
-
-@Builder
-public record LoginResponse(String jwtToken, boolean ifFirstLogin) {
-}
