+++ src/main/java/com/api/ttoklip/global/security/oauth2/dto/OAuthLoginRequest.java
@@ -1,4 +1,4 @@
-package com.api.ttoklip.global.security.auth.dto;
+package com.api.ttoklip.global.security.oauth2.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class LoginRequest {
+public class OAuthLoginRequest {
 
     @Schema(type = "string", description = "oauth accessToken", example = "AAAAux5O0y30x7G1twup/hPQIdsu/B3i3WL490lghVyU=")
     private String accessToken;
