+++ src/main/java/com/api/ttoklip/global/security/oauth2/dto/OAuthLoginResponse.java
@@ -1,0 +1,7 @@
+package com.api.ttoklip.global.security.oauth2.dto;
+
+import lombok.Builder;
+
+@Builder
+public record OAuthLoginResponse(String jwtToken, boolean ifFirstLogin) {
+}
