+++ src/main/java/com/api/ttoklip/global/security/oauth2/userInfo/KakaoUserInfo.java
@@ -1,4 +1,4 @@
-package com.api.ttoklip.global.security.auth.userInfo;
+package com.api.ttoklip.global.security.oauth2.userInfo;
 
 import com.api.ttoklip.global.exception.ApiException;
 import com.api.ttoklip.global.exception.ErrorType;
